// src/app/shared/enums/sorting.enum.ts
export enum Sorting {
  ASC = 'asc',
  DESC = 'desc',
}

export enum OrderBy {
  ASC = 1,
  DESC = -1,
}

export enum SortBy {
  TEAM_MEMBER = 'name.first',
  EMAIL = 'email',
  ROLE = 'role.name',
  NAME = 'name',
  SERVICE_REPRESENTATIVE = 'serviceRepresentative',
  SERVICE_REPRESENTATIVE_NAME = 'serviceRepresentative.name'
}